backend/products_store/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
django-filter==25.1
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
